requirements.txt
# ============================================================================
# CareOps — Python Dependencies
# ============================================================================

# -- Core Framework --
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12

# -- Supabase --
supabase>=2.10.0

# -- HTTP Client --
httpx>=0.27.0

# -- Retry/Resilience --
tenacity>=9.0.0

# -- Google Cloud --
google-cloud-texttospeech>=2.17.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0

# -- Validation --
email-validator>=2.2.0

# -- Dev/Testing --
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
